Close #114.
The new environment is a multi-segment stack where a segment is  created in a lazy manner for each world. It results in small (but numerous) arrays that are fast to expand.

This required to backport a lot of code in Choco due to a bad "final" keyword. Once choco 4 released, the code will be cleaned.