Fix #118.
The list of VMs and nodes parsed in the past are put into a cache to deduplicate as possible.

For simplicity, the helper only provides list of elements as outputs. This required to move to sets from times to times at the converter levels but it increases the cache hit ratio.