Simplify the simulator.
Temporary states are mis-leading and provide a precision that might be
dangerous once we move to an event based agenda. So basically, we start
to remove some temporary states like VM_HALTING when the shutdown action
starts. Now, the VM state is changed only once the action terminates
which sounds safer