Close #313.
The initialisation of Knapsack checks the number of open variables to avoid doing changes in the bitsets.

With a few open variables (typically in repair mode), it starts
with an empty bitset as there will be a few candidate. With numerous open variables (so many candidates), it starts with bitsets being sets.

This reduces drastically the time spent in postInitialise() for large repair-based problems.