Initial slot based reasoning.
VectorPacking constructor is updated to provide a fastpath for
a 1-D packing problem while the complete contructor can indicate
that the last dimension is devoted to the cardinality dimension.

We start addressing #307 with that new constructor. For the moment,
this is done only during the initial propagation. Given the item
sizes, we cap the cardinality dimension.