TIntMap structure to replace TIntObjectMap.
This structure will be used to replace TIntObjectMap when
being using to associate an integer to an Entity object.
The structure is faster for put() as there are no key space
compaction while providing an efficient copy() method for
deep copy.

This is already applied on ShareableResource.
setConsumption() or setCapacity() no longer allocate
if the backend table is already sized properly while
copy() is now mostly free (minor allocations, negligible
CPU time).

The iteration over VMs and Nodes in this view is now
deprecated because it is inefficient.