New ReconfigurationProblem#makeHostVariable.
This helper creates a host variable with a known domain.
This will help creating right-sized domains without having
to use a Fence or a Ban constraint to restrict it in a later
stage.

Furthermore, one minor optimisation in InstanceSolverRunner
to reduce the time spent in verifying the nodes involved
in constraints.