Close issue #382.
This fixes an overfiltering in VectorPackingPropagator.
In case of backtrack, the events related to load assignment may
be queued prior to those related to item assigments. In that case,
the knapsack constraint was able to deduce the item assignment from
the loads update. Then, when the legacy item assignment event is
handled, the loads LB was increased again.

Now, we consider that the bitsets in the knapsack constraint is
also the source of truth so in case assignItem it called, we check
early if the item was still candidate for the assigned bin to avoid
the double counting.