Remove useless Variables.
Resource allocation per VM was stored inside a choco variable.
This was useless given eventually we expect a constant in our
constraints.

This reduces the number of constants in the model. In practice,
this is not a game changer given constants are deduplicated
by choco and ignores when they are not attached to any propagator.

Close #424