Isolate and aggregate constraints to model staying VMs.
This used to be a collection of per-VM constraints:
- 3 to model the action duration
- 1 for the symmetry breaking

This impacted the scalability of the model as the number of
constraints grows with the number of running VMs. Eventually,
most of the model was made of those small constraints.

Now, everything is packaged inside a single propagator that
embeds the same logic. This reduces the number of propagators
while reducing a bit the memory usage.

Close #425.