Provides misplacedVMs from ReconfigurationProblem.
This set of VMs is computed from the views and the constraints.
In repair mode, it is used to restrict the problem by fixing
some VMs but in the rebuild mode, this set was ignored and not
available.

In practice, it is however something useful to have inside the
search heuristics to help focusing on critical VMs.

Close #491.